doh-data-ingest/requirements.txt
requests==2.23.0
beautifulsoup4==4.8.2
boto3==1.12.32
pendulum==2.1.0
websocket-client==0.57.0
